refactor: streamline recovery logic by introducing normalizeRecoveryPlan and revising protected regions
